Add call to action section
Added a call to action section with some text and a button. Not sure I
am completely happy with how it looks/acts (mostly when the screen gets
smaller), but I'm also not sure how to improve it at this point!

It mostly looks like how the example image from TOP looks, so I think
it is good enough for now!